修改公司管理数据类型

+++ visa/src/main/java/io/znz/jsite/visa/entity/company/CompanyEntity.java
@@ -29,7 +29,7 @@ public class CompanyEntity {
 	private Integer id;
 	@Column
 	@Comment("管理员账号id")
-	private Integer adminId;
+	private long adminId;
 	@Column
 	@Comment("管理员姓名")
 	private String adminName;
